+++ commands.c
@@ -1,0 +1,27 @@
+/****************************************************************
+ * Author: Elizabeth Acevedo
+ * 
+ * Created: 02/03/2025
+ * 
+ * File: commands.c
+ * 
+ * Description: This program 
+ * 
+ * Utilization: 
+ * 
+ * 
+ ****************************************************************/
+/* Includes */
+#include "commands.h"
+
+/* Definitions */
+
+/* Function Prototypes */
+void sendMsgCmd(uint8_t *dataBuffer);
+
+void sendMsgCmd(uint8_t *dataBuffer) {
+    
+    
+    // Send the message to the server
+    //sendPDU(socketNum, dataBuffer, sendLen);
+}